+++ railties/test/application/rake/dbs_test.rb
@@ -488,19 +488,30 @@ module ApplicationTests
         end
       end
 
+      ["dump", "load"].each do |command|
+        test "db:structure:#{command} is deprecated" do
+          add_to_config("config.active_support.deprecation = :stderr")
+          stderr_output = capture(:stderr) { rails("db:structure:#{command}", stderr: true, allow_failure: true) }
+          assert_match(/DEPRECATION WARNING: Using `bin\/rails db:structure:#{command}` is deprecated and will be removed in Rails 6.2/, stderr_output)
+        end
+      end
+
       test "db:structure:dump and db:structure:load without database_url" do
+        add_to_config "config.active_record.schema_format = :sql"
         require "#{app_path}/config/environment"
         db_config = ActiveRecord::Base.connection_db_config
         db_structure_dump_and_load db_config.database
       end
 
       test "db:structure:dump and db:structure:load with database_url" do
+        add_to_config "config.active_record.schema_format = :sql"
         require "#{app_path}/config/environment"
         set_database_url
         db_structure_dump_and_load database_url_db_name
       end
 
       test "db:structure:dump and db:structure:load set ar_internal_metadata" do
+        add_to_config "config.active_record.schema_format = :sql"
         require "#{app_path}/config/environment"
         db_config = ActiveRecord::Base.connection_db_config
         db_structure_dump_and_load db_config.database
@@ -510,6 +521,7 @@ module ApplicationTests
       end
 
       test "db:structure:dump does not dump schema information when no migrations are used" do
+        add_to_config "config.active_record.schema_format = :sql"
         # create table without migrations
         rails "runner", "ActiveRecord::Base.connection.create_table(:posts) {|t| t.string :title }"
 
@@ -534,6 +546,7 @@ module ApplicationTests
         rails "db:schema:load"
         assert_equal '["posts", "comments", "schema_migrations", "ar_internal_metadata"]', list_tables[]
 
+        add_to_config "config.active_record.schema_format = :sql"
         app_file "db/structure.sql", <<-SQL
           CREATE TABLE "users" ("id" INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, "name" varchar(255));
         SQL
@@ -588,8 +601,15 @@ module ApplicationTests
       end
 
       test "db:test:load_structure without database_url" do
+        add_to_config "config.active_record.schema_format = :sql"
         require "#{app_path}/config/environment"
         db_test_load_structure
+      end
+
+      test "db:test:load_structure is deprecated" do
+        add_to_config("config.active_support.deprecation = :stderr")
+        stderr_output = capture(:stderr) { rails("db:test:load_structure", stderr: true, allow_failure: true) }
+        assert_match(/DEPRECATION WARNING: Using `bin\/rails db:test:load_structure` is deprecated and will be removed in Rails 6.2/, stderr_output)
       end
 
       test "db:setup loads schema and seeds database" do